code/requirements.txt
discord.py==1.3.3
scrapy==2.1.0
